move helpers out of test and into test_util.  Add some doc strings
