+++ tests/parkfield/make_processing_configs.py
@@ -6,8 +6,8 @@ from aurora.general_helper_functions import BAND_SETUP_PATH
 from aurora.config.config_creator import ConfigCreator
 from aurora.config.decimation_level_config import DecimationLevelConfig
 
-from helpers import DATA_PATH
-from helpers import CONFIG_PATH
+from aurora.test_utils.parkfield.path_helpers import DATA_PATH
+from aurora.test_utils.parkfield.path_helpers import CONFIG_PATH
 
 CHANNEL_SCALE_FACTORS = {}
 CHANNEL_SCALE_FACTORS["PKD"] = {}
