+++ tests/parkfield/test_calibrate_parkfield.py
@@ -2,10 +2,43 @@ from aurora.general_helper_functions import TEST_PATH
 from aurora.time_series.windowing_scheme import WindowingScheme
 from mth5.mth5 import MTH5
 from aurora.test_utils.parkfield.calibration_helpers import parkfield_sanity_check
-from helpers import AURORA_RESULTS_PATH
+from aurora.test_utils.parkfield.path_helpers import AURORA_RESULTS_PATH
+
 
 
 def validate_bulk_spectra_have_correct_units(run_obj, run_ts_obj, show_spectra=False):
+    """
+
+    Parameters
+    ----------
+    run_obj: mth5.groups.master_station_run_channel.RunGroup
+        /Survey/Stations/PKD/001:
+        ====================
+            --> Dataset: ex
+            .................
+            --> Dataset: ey
+            .................
+            --> Dataset: hx
+            .................
+            --> Dataset: hy
+            .................
+            --> Dataset: hz
+            .................
+    run_ts_obj: mth5.timeseries.run_ts.RunTS
+        RunTS Summary:
+            Station:     PKD
+            Run:         001
+            Start:       2004-09-28T00:00:00+00:00
+            End:         2004-09-28T01:59:59.950000+00:00
+            Sample Rate: 40.0
+            Components:  ['ex', 'ey', 'hx', 'hy', 'hz']
+    show_spectra: bool
+        controls whether plots flash to screen in parkfield_sanity_check
+
+    Returns
+    -------
+
+    """
     windowing_scheme = WindowingScheme(
         taper_family="hamming",
         num_samples_window=run_ts_obj.dataset.time.shape[0],  # 288000
@@ -26,6 +59,7 @@ def validate_bulk_spectra_have_correct_units(run_obj, run_ts_obj, show_spectra=F
         show_spectra=show_spectra,
         include_decimation=True,
     )
+    return
 
 
 def test():
