+++ tests/parkfield/test_process_parkfield_run.py
@@ -4,11 +4,11 @@ import pickle
 from aurora.config.processing_config import RunConfig
 
 from aurora.pipelines.process_mth5 import process_mth5_run
+from aurora.test_utils.parkfield.path_helpers import AURORA_RESULTS_PATH
+from aurora.test_utils.parkfield.path_helpers import EMTF_RESULTS_PATH
 from aurora.transfer_function.plot.comparison_plots import compare_two_z_files
 
 
-from helpers import AURORA_RESULTS_PATH
-from helpers import EMTF_RESULTS_PATH
 from make_parkfield_mth5 import test_make_parkfield_mth5
 from make_processing_configs import create_run_test_config
