+++ tests/parkfield/test_process_parkfield_run_rr.py
@@ -7,8 +7,8 @@ from aurora.transfer_function.plot.comparison_plots import compare_two_z_files
 from make_parkfield_mth5 import test_make_parkfield_hollister_mth5
 from make_processing_configs import create_run_test_config_remote_reference
 
-from helpers import AURORA_RESULTS_PATH
-from helpers import EMTF_RESULTS_PATH
+from aurora.test_utils.parkfield.path_helpers import AURORA_RESULTS_PATH
+from aurora.test_utils.parkfield.path_helpers import EMTF_RESULTS_PATH
 
 
 def test_processing(z_file_path=None):
